Incremental rebuild, LAN-accessible search API, UI polish
Add rebuild.py for incremental ingestion (hash-based change detection,
only re-embeds modified episodes). Add sync.sh for rsyncing episode data.
Bind search API to 0.0.0.0 so Caddy can proxy from LAN clients. Fix Caddy
Ollama proxy headers, support fractional episode numbers, update UI with
about section and sample queries. Update README for current project state.

+++ ingest/build_db.py
@@ -27,7 +27,7 @@ def create_schema(conn, embedding_dim):
     conn.executescript("""
         CREATE TABLE IF NOT EXISTS episodes (
             id INTEGER PRIMARY KEY,
-            number INTEGER UNIQUE NOT NULL,
+            number REAL UNIQUE NOT NULL,
             title TEXT NOT NULL,
             pub_date TEXT,
             synopsis TEXT,
